test: add accept offers test case
